requirements.txt
captum == 0.4.1
matplotlib == 3.4.0
numpy == 1.22.3
optuna == 2.10.0
pandas == 1.3.4
pytorch == 1.10.1
pytorch-lightning == 1.5.9
scipy == 1.7.3
seaborn == 0.11.2
skimage == 0.19.3
sklearn == 1.0.2
torchmetrics == 0.5.0
tslearn == 0.5.2

